requirements.txt
asgiref==3.7.2
black==23.9.1
boto3==1.28.54
celery==5.3.4
Django==4.2.5
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-jazzmin==2.6.0
django-json-widget==1.1.1
django-redis-cache==3.0.1
django-rest-passwordreset==1.3.0
django-storages==1.14
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg2==1.19.4
ipython==8.15.0
isort==5.12.0
pre-commit==3.4.0
psycopg2==2.9.9
python-decouple==3.8
redis==3.5.3
requests==2.31.0
shortuuid==1.0.11
typing_extensions==4.8.0